% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASMC2_LP2.R
\name{ASMC2_LP2}
\alias{ASMC2_LP2}
\title{Annealed sequential Monte Carlo}
\usage{
ASMC2_LP2(
  K,
  smc_tuning_param,
  data,
  is_unknownPar,
  ODEmodel,
  ODEparameters,
  likelihood,
  likeliParam_fixed,
  likeliParam_unknown = NULL,
  reference_prior_list,
  hyperpar,
  n_core = 1
)
}
\arguments{
\item{K}{Number of particles}

\item{smc_tuning_param}{List of parameters for the SMC algorithm. Must contain eps and phi.}

\item{data}{Observed data}

\item{ODEmodel}{Function defining the ODE model. Must be compatible with deSolve functions.}

\item{ODEparameters}{Named vector of DE parameters}

\item{likelihood}{List containing density/mass of observations}

\item{likeliParam_fixed}{Additional parameters to be passed to the likelihood which are known}

\item{likeliParam_unknown}{Additional parameters of the likelihood which are to be estimated. NULL if none applicable.}

\item{reference_prior_list}{List containing reference & prior distributions}

\item{hyperpar}{List of hyperparameters. Must contain sigma_theta1, sigma_theta2, and ksi}

\item{n_core}{Number of cores to be used for parallel processing. Default value is 1.}

\item{is_unknowPar}{Boolean vector indicating which parameters are being estimated}
}
\value{
List of results
}
\description{
Annealed SMC for estimating DE parameters & likelihood parameters
}
