% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginalLikelihood.R
\name{getMarginalLikelihood}
\alias{getMarginalLikelihood}
\title{Estimate marginal likelihood from MCMC samples}
\usage{
getMarginalLikelihood(
  result_list,
  burn_in,
  is_unknownPar,
  data,
  ODEmodel,
  likelihood,
  likeliParam_fixed = NULL,
  hyperpar,
  prior_list,
  lower_bound,
  upper_bound,
  ...
)
}
\arguments{
\item{result_list}{List of same format as the sample_list compoenent of the MCMC_SIR output}

\item{burn_in}{Number of samples to use as burn_in}

\item{is_unknownPar}{Boolean vector indicating which DE parameters are to be estimated}

\item{data}{List of data}

\item{ODEmodel}{ODE model}

\item{likelihood}{List containing likelihood density/mass}

\item{likeliParam_fixed}{Fixed parameters to be passed to the likelihood. Leave as NULL if all additional parameters are estimated.}

\item{hyperpar}{List of hyperparameters; same as was used to generate the MCMC output}

\item{prior_list}{List of prior densities for theta and basis function coefficients}

\item{lower_bound}{Named vector of lower bounds on estimated parameters to be passed to bridge_sampler}

\item{upper_bound}{Named vector of upper bounds on estimated paramters to be passed to bridge_sampler}

\item{...}{Additional parameters to be passed to bridge_sampler}

\item{basis_funs}{List of basis function expansion components matching basisFuns output}
}
\value{
Estimated marginal likelihood of the data
}
\description{
Uses bridge sampling to estimate the marginal likelihood of the data from the MCMC samples. Uses the bridgesampling package.
}
