% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC2_likeliparam.R
\name{MCMC2_LP}
\alias{MCMC2_LP}
\title{MCMC for SIR model}
\usage{
MCMC2_LP(
  K,
  data,
  is_unknownPar,
  ODEmodel,
  init_DEpar,
  likelihood,
  likeliParam_fixed,
  init_likeliParam_unknown = NULL,
  prior_list,
  hyperpar
)
}
\arguments{
\item{K}{Number of samples to generate}

\item{data}{List of observed data}

\item{is_unknownPar}{Boolean vector indicating which parameters are to be estimated}

\item{ODEmodel}{Function defining the DE system. Must be compatible with deSolve.}

\item{init_DEpar}{Initial estimates for the DE parameters}

\item{likelihood}{List containing funcitons for simulating from and evaluating the likelihood of the data.}

\item{likeliParam_fixed}{Vector of fixed likelihood parameters.}

\item{init_likeliParam_unknown}{Vector of initial values for the unknown likeliood parameters.}

\item{prior_list}{List of prior distributions and parameters}

\item{hyperpar}{List of hyperparameters}
}
\value{
List containing posterior distribution of DE parameters and basis coefficients
}
\description{
Markov chain Monte Carlo for estimating parameters of susceptible-infected-recovered model with or without additional likelihood parameters.
}
